+++ src/test/resources/org/csanchez/jenkins/plugins/kubernetes/pipeline/declarative.groovy
@@ -16,11 +16,20 @@ pipeline {
   stages {
     stage('Run maven') {
       steps {
+        sh 'set'
+        sh "echo OUTSIDE_CONTAINER_ENV_VAR = ${CONTAINER_ENV_VAR}"
         container('maven') {
           sh 'echo INSIDE_CONTAINER_ENV_VAR = ${CONTAINER_ENV_VAR}'
           sh 'mvn -version'
         }
       }
     }
+	stage('Run maven with a different shell') {
+		steps {
+		  container(name: 'maven', shell: '/bin/bash') {
+			sh 'mvn -version'
+		  }
+		}
+	  }
   }
 }